+++ HISTORY.md
@@ -1,7 +1,7 @@
 # History
 
 
-# 2018-04-08, version 5.9.0
+# 2019-04-08, version 5.9.0
 
 - Implemented functions `row` and `column` (see #1413). Thanks @SzechuanSage.
 - Fixed #1459: `engineering` notation of function `format` not available